se añadio el tipo double al hast table freq

+++ HashTableFreq.h
@@ -17,7 +17,7 @@ Node * getHashTableFreqNode(HashTableFreq * self,  Node * node);
 void printHashTableFreq(HashTableFreq * self);
 
 int getTotalCharsCounted(HashTableFreq * self);
-int getFrequencyById(HashTableFreq * self, int index);
+double getFrequencyById(HashTableFreq * self, int index);
 char getCharById(HashTableFreq * self, int index);
 
 #endif
